modify existing tests
